Refactor sync committee tests so rewards are verified for all test cases
